fix(db): cast sql<number> money aggregates to ::float8 for honest types (#524)
Drizzle numeric(12,2) columns come back from postgres.js as strings, but
several sql<number>/db.execute<{...: number}> sites over total_amount,
unit_price, total_price, and monthly_budget claimed a real number without
casting, matching #477's already-fixed pattern in alerts.ts. None were live
bugs (every consumer wrapped in Number(...) or hit arithmetic operators that
coerce), but the annotation was an unchecked lie a future edit could exploit.

Fixed sites: trend.ts, supplier-list-query.ts, suppliers/+page.server.ts,
dashboard/+page.server.ts, chat-context.ts — all cast to ::float8 (display/
threshold aggregates, per the issue's decision rule). supplier-reliability.ts
switches to string + moneyToNumber() since those unit prices are re-averaged
in JS. Every other sql/execute site touching the six money-column families
was already honest (string annotation + money.ts, or an existing ::float8/
::int cast) after a full grep-based re-inventory.

Adds a source-scan test asserting every sql<number> site referencing a money
column carries a numeric cast, plus a DB-backed spot check showing the
uncast/cast difference against real Postgres.

+++ src/lib/server/supplier-list-query.ts
@@ -12,7 +12,7 @@ export function supplierCategoryExpr(): SQL<string> {
 }
 
 export function supplierTotalSpendExpr(): SQL<number> {
-	return sql<number>`COALESCE(SUM(COALESCE(${invoices.totalAmount}, 0)), 0)`;
+	return sql<number>`COALESCE(SUM(COALESCE(${invoices.totalAmount}, 0)), 0)::float8`;
 }
 
 export function supplierLastInvoiceExpr(): SQL<string | null> {
