+++ src/routes/(app)/suppliers/+page.server.ts
@@ -48,14 +48,14 @@ export const load: PageServerLoad = async ({ url, locals }) => {
 				createdAt:        suppliers.createdAt,
 				category:         supplierCategoryExpr().as('category'),
 				total_spend:      supplierTotalSpendExpr().as('total_spend'),
-				month_spend:      sql<number>`COALESCE(SUM(CASE WHEN TO_CHAR(${invoices.invoiceDate},'YYYY-MM')=TO_CHAR(NOW(),'YYYY-MM') THEN COALESCE(${invoices.totalAmount},0) ELSE 0 END),0)`.as('month_spend'),
+				month_spend:      sql<number>`COALESCE(SUM(CASE WHEN TO_CHAR(${invoices.invoiceDate},'YYYY-MM')=TO_CHAR(NOW(),'YYYY-MM') THEN COALESCE(${invoices.totalAmount},0) ELSE 0 END),0)::float8`.as('month_spend'),
 				open_count:       sql<number>`COUNT(CASE WHEN ${invoices.status}='pending' THEN 1 END)`.as('open_count'),
 				invoice_count:    sql<number>`COUNT(${invoices.id})`.as('invoice_count'),
 				last_invoice_date:   sql<string | null>`MAX(${invoices.invoiceDate})`.as('last_invoice_date'),
 				has_overdue:         sql<number>`MAX(CASE WHEN ${invoices.status}='pending' AND ${invoices.dueDate} IS NOT NULL AND ${invoices.dueDate} < ${today} THEN 1 ELSE 0 END)`.as('has_overdue'),
 				has_due_soon:        sql<number>`MAX(CASE WHEN ${invoices.status}='pending' AND ${invoices.dueDate} IS NOT NULL AND ${invoices.dueDate} BETWEEN ${today} AND ${weekEnd} THEN 1 ELSE 0 END)`.as('has_due_soon'),
 				month_invoice_count: sql<number>`COALESCE(COUNT(CASE WHEN TO_CHAR(${invoices.invoiceDate},'YYYY-MM')=TO_CHAR(NOW(),'YYYY-MM') THEN 1 END),0)`.as('month_invoice_count'),
-				last_month_spend:    sql<number>`COALESCE(SUM(CASE WHEN TO_CHAR(${invoices.invoiceDate},'YYYY-MM')=TO_CHAR(NOW()-INTERVAL'1 month','YYYY-MM') THEN COALESCE(${invoices.totalAmount},0) ELSE 0 END),0)`.as('last_month_spend'),
+				last_month_spend:    sql<number>`COALESCE(SUM(CASE WHEN TO_CHAR(${invoices.invoiceDate},'YYYY-MM')=TO_CHAR(NOW()-INTERVAL'1 month','YYYY-MM') THEN COALESCE(${invoices.totalAmount},0) ELSE 0 END),0)::float8`.as('last_month_spend'),
 			})
 				.from(suppliers)
 				.leftJoin(invoices, and(eq(invoices.supplierId, suppliers.id), tdb.scope(invoices.restaurantId)))
@@ -71,7 +71,7 @@ export const load: PageServerLoad = async ({ url, locals }) => {
 				SELECT
 					i.supplier_id,
 					TO_CHAR(i.invoice_date, 'YYYY-MM') AS month,
-					AVG(ili.unit_price) AS avg_price
+					AVG(ili.unit_price)::float8 AS avg_price
 				FROM invoice_line_items ili
 				JOIN invoices i ON i.id = ili.invoice_id
 				WHERE ili.unit_price IS NOT NULL
